requirements.txt
pandas==0.23.4
GitPython==2.1.11
